pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fleet-mem"
version = "0.1.0"
description = "Shared code intelligence for agent fleets — AST-aware semantic search + multi-agent memory with git-concurrent coordination"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["mcp", "agent-memory", "code-search", "multi-agent", "semantic-search", "tree-sitter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp>=1.0.0",
    "chromadb>=0.5.0",
    "ollama>=0.3.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "python-dotenv>=1.0.0",
    "xxhash>=3.4.0",
    "watchdog>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/sam-ent/fleet-mem"
Repository = "https://github.com/sam-ent/fleet-mem"
Issues = "https://github.com/sam-ent/fleet-mem/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4.0",
]
languages = [
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-java>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["fleet_mem*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
